feat: update Files screen
